creacion de una ruta para todas las artes y filtrado por tipo y/o artista
