+++ src/server.js
@@ -4,21 +4,17 @@ const cors = require('cors')
 const artistRouter = require('./routes/artist')
 const authRouter = require('./routes/auth')
 const userRouter = require('./routes/user')
-const muralRouter = require('./routes/mural')
-const graffitiRouter = require('./routes/graffiti')
-const stickerRouter = require('./routes/sticker')
+const streetartRouter = require('./routes/streetart')
 
 const app = express()
 
-app.use(cors())
 app.use(express.json())
+app.use(cors())
 
 app.use('/auth', authRouter)
 app.use('/artist', artistRouter)
 app.use('/user', userRouter)
-app.use('/mural', muralRouter)
-app.use('/graffiti', graffitiRouter)
-app.use('/sticker', stickerRouter)
+app.use('/streetart', streetartRouter)
 
 app.get('/', (request, response) => {
     response.json({
